improved file layout
Reworked the file layout to more accurately represent the underlying
algorithms in each file as well as divided the tests into the "edit" and
"overlap" folders. Added a few new algorithms (Hamann, SMC, length
ratio) and will add two other simple overlap algorithms such as prefix
and postfix similarity
