+++ CITATION.cff
@@ -5,6 +5,6 @@ authors:
   given-names: "Andrew Marcus Ryan"
   orcid: "https://orcid.org/0009-0000-5260-3047"
 title: "Goombay"
-version: 0.15.0
-date-released: 2025-02-20
+version: 0.21.0
+date-released: 2025-06-20
 url: "https://github.com/lignum-vitae/goombay"
