+++ tests/test_edit/test_Hamming.py
