+++ tests/test_edit/test_SmithWaterman.py
