+++ tests/test_overlap/test_LCString.py
